help

+++ containers/rencos-env/rencos-master/onmt/translate/translator.py
@@ -585,7 +585,7 @@ class Translator(object):
             length_penalty = ((5.0 + (step + 1)) / 6.0) ** alpha
 
             # Flatten probs into a list of possibilities.
-            curr_scores = (log_probs / length_penalty).long()
+            curr_scores = (log_probs / length_penalty)
             curr_scores = curr_scores.reshape(-1, beam_size * vocab_size)
             topk_scores, topk_ids = curr_scores.topk(beam_size, dim=-1)
 
@@ -606,7 +606,7 @@ class Translator(object):
 
             # Append last prediction.
             alive_seq = torch.cat(
-                [alive_seq.index_select(0, select_indices),
+                [alive_seq.index_select(0, select_indices.long()),
                  topk_ids.contiguous().view(-1, 1)], -1)
             if return_attention:
                 current_attn = attn.index_select(1, select_indices)
